requirements.txt
numpy>=1.23
pandas>=1.5
matplotlib>=3.6
Pillow>=9.0

torch>=2.0
torchvision>=0.15
torchcam>=0.4

scikit-learn>=1.2
shap>=0.41
tqdm>=4.65

streamlit>=1.25
opencv-python>=4.7
